Split resources routes into separate module
