+++ app/api/routes/__init__.py
@@ -1,1 +1,2 @@
 from app.api.routes import old_routes
+from app.api.routes import resources
